+++ client/src/components/FeedbackForms.js
@@ -1,181 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  FormControl,
   FormControlLabel,
   FormLabel,
-  FormHelperText,
   Radio,
   RadioGroup,
   TextField,
-  Select,
 } from '@material-ui/core/';
 import { useStyles } from '../themes/theme';
-import { useHistory } from 'react-router-dom';
-import { NextQuestionButton, PreviousQuestionButton } from '../components/Buttons';
-import { createFeedback, getFeedbackCreator } from '../utils/apiFunctions';
-
+import { getFeedbackCreator } from '../utils/apiFunctions';
+import { NextPage, PrevPage } from './FeedbackButtons';
 const interviewId = '5fbfcab4d54a3f7654ae5684'; // For testing purposes. To be changed later to pull from params
-
-const NextPage = ({
-  feedbackType,
-  performanceLevel,
-  categories,
-  strengths,
-  improvements,
-  resources,
-  other,
-}) => {
-  const classes = useStyles();
-  const history = useHistory();
-  const [error, setError] = useState(false);
-
-  const addPerformance = async () => {
-    if (performanceLevel) {
-      try {
-        await createFeedback(interviewId, { performanceLevel: performanceLevel });
-        history.push(`/interview/feedback/2`);
-      } catch (error) {
-        console.error(error);
-      }
-    } else {
-      setError(true);
-    }
-  };
-
-  const addCategories = async () => {
-    const {
-      communication,
-      codeEfficiency,
-      codeOrganization,
-      speed,
-      debugging,
-      problemSolving,
-    } = categories;
-
-    if (
-      communication &&
-      codeEfficiency &&
-      codeOrganization &&
-      speed &&
-      debugging &&
-      problemSolving
-    ) {
-      try {
-        await createFeedback(interviewId, {
-          categories: categories,
-        });
-        history.push(`/interview/feedback/3`);
-      } catch (error) {
-        console.error(error);
-      }
-    } else {
-      setError(true);
-    }
-  };
-
-  const addStrengths = async () => {
-    if (strengths) {
-      try {
-        await createFeedback(interviewId, { strengths: strengths });
-        history.push(`/interview/feedback/4`);
-      } catch (error) {
-        console.error(error);
-      }
-    } else {
-      setError(true);
-    }
-  };
-
-  const addImprovements = async () => {
-    if (improvements) {
-      try {
-        await createFeedback(interviewId, {
-          improvements: improvements,
-        });
-        history.push(`/interview/feedback/5`);
-      } catch (error) {
-        console.error(error);
-      }
-    } else {
-      setError(true);
-    }
-  };
-
-  const addResources = async () => {
-    if (resources) {
-      try {
-        await createFeedback(interviewId, {
-          resources: resources,
-        });
-        history.push(`/interview/feedback/6`);
-      } catch (error) {
-        console.error(error);
-      }
-    } else {
-      setError(true);
-    }
-  };
-
-  const addOther = async () => {
-    if (other) {
-      try {
-        await createFeedback(interviewId, {
-          other: other,
-        });
-        history.push(`/dashboard`);
-      } catch (error) {
-        console.error(error);
-      }
-    } else {
-      setError(true);
-    }
-  };
-
-  const nextPage = async (e) => {
-    e.preventDefault();
-
-    switch (feedbackType) {
-      case 'performance':
-        addPerformance();
-        break;
-      case 'categories':
-        addCategories();
-        break;
-      case 'strengths':
-        addStrengths();
-        break;
-      case 'improvements':
-        addImprovements();
-        break;
-      case 'resources':
-        addResources();
-        break;
-      case 'other':
-        addOther();
-        break;
-    }
-  };
-
-  return (
-    <>
-      {error ? <FormHelperText>Please select a rating</FormHelperText> : ''}
-      <NextQuestionButton
-        variant='outlined'
-        onClick={nextPage}
-        color='primary'
-        autoFocus
-      >
-        NEXT QUESTION
-      </NextQuestionButton>
-    </>
-  );
-};
 
 export function FormOne() {
   const classes = useStyles();
@@ -223,12 +57,14 @@ export function FormOne() {
         </RadioGroup>
         <FormLabel className={classes.feedbackFormLabel}>Perfect</FormLabel>
       </div>
-      <NextPage feedbackType='performance' performanceLevel={performanceLevel} />
+      <div>
+        <NextPage feedbackType='performance' performanceLevel={performanceLevel} />
+      </div>
     </>
   );
 }
 
-export function FormTwo() {
+export function FormTwo({ pageNumber }) {
   const classes = useStyles();
   const [categoryRatings, setCategories] = useState({});
 
@@ -361,12 +197,15 @@ export function FormTwo() {
           })}
         </div>
       </div>
-      <NextPage feedbackType='categories' categories={categoryRatings} />
+      <div>
+        <PrevPage pageNumber={pageNumber} />
+        <NextPage feedbackType='categories' categories={categoryRatings} />
+      </div>
     </>
   );
 }
 
-export function FormThree() {
+export function FormThree({ pageNumber }) {
   const classes = useStyles();
   const [strengths, setStrength] = useState();
 
@@ -395,12 +234,15 @@ export function FormThree() {
         onChange={changeStrengths}
         defaultValue={strengths}
       />
-      <NextPage feedbackType='strengths' strengths={strengths} />
+      <div>
+        <PrevPage pageNumber={pageNumber} />
+        <NextPage feedbackType='strengths' strengths={strengths} />
+      </div>
     </>
   );
 }
 
-export function FormFour() {
+export function FormFour({ pageNumber }) {
   const classes = useStyles();
   const [improvements, setImprovements] = useState();
 
@@ -429,12 +271,15 @@ export function FormFour() {
         onChange={changeImprovements}
         defaultValue={improvements}
       />
-      <NextPage feedbackType='improvements' improvements={improvements} />
+      <div>
+        <PrevPage pageNumber={pageNumber} />
+        <NextPage feedbackType='improvements' improvements={improvements} />
+      </div>
     </>
   );
 }
 
-export function FormFive() {
+export function FormFive({ pageNumber }) {
   const classes = useStyles();
   const [resources, setResources] = useState();
 
@@ -463,12 +308,15 @@ export function FormFive() {
         onChange={changeResources}
         defaultValue={resources}
       />
-      <NextPage feedbackType='resources' resources={resources} />
+      <div>
+        <PrevPage pageNumber={pageNumber} />
+        <NextPage feedbackType='resources' resources={resources} />
+      </div>
     </>
   );
 }
 
-export function FormSix() {
+export function FormSix({ pageNumber }) {
   const classes = useStyles();
   const [other, setOther] = useState();
 
@@ -497,7 +345,10 @@ export function FormSix() {
         onChange={changeOther}
         defaultValue={other}
       />
-      <NextPage feedbackType='other' other={other} />
+      <div>
+        <PrevPage pageNumber={pageNumber} />
+        <NextPage feedbackType='other' other={other} />
+      </div>
     </>
   );
 }
